+++ tests/utils/test_graph.py
@@ -130,3 +130,16 @@ def test_download_graph__add_elevation() -> None:
     )
     nodes, _ = osmnx.utils_graph.graph_to_gdfs(graph)
     assert "elevation" in nodes.columns.tolist()
+
+
+@pytest.mark.default_cassette("download_nailsea.yaml")
+@pytest.mark.vcr("download_nailsea_elevation.yaml")
+def test_download_graph__add_grades() -> None:
+    """
+    Make sure grades are attached once an Google API key has been provided.
+    """
+    graph = graph_utils.download_graph(
+        "Nailsea, North Somerset, England", api_key="some_random_api_key"
+    )
+    _, edges = osmnx.utils_graph.graph_to_gdfs(graph)
+    assert "grade" in edges.columns.tolist()